Comment reviews about renamings
