+++ microprofile/testing/junit5/src/main/java/io/helidon/microprofile/testing/junit5/HelidonJunitExtension.java
@@ -36,7 +36,7 @@ import io.helidon.config.mp.MpConfigSources;
 import io.helidon.config.yaml.mp.YamlMpConfigSource;
 import io.helidon.microprofile.server.JaxRsCdiExtension;
 import io.helidon.microprofile.server.ServerCdiExtension;
-import io.helidon.microprofile.testing.junit5.ConfigBlob.Type;
+import io.helidon.microprofile.testing.junit5.AddConfigBlock.Type;
 
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.enterprise.context.Dependent;
@@ -118,7 +118,7 @@ class HelidonJunitExtension implements BeforeAllCallback,
         AddConfig[] configs = getAnnotations(testClass, AddConfig.class);
         classLevelConfigMeta.addConfig(configs);
         classLevelConfigMeta.configuration(testClass.getAnnotation(Configuration.class));
-        classLevelConfigMeta.configBlob(testClass.getAnnotation(ConfigBlob.class));
+        classLevelConfigMeta.addConfigBlock(testClass.getAnnotation(AddConfigBlock.class));
         configProviderResolver = ConfigProviderResolver.instance();
 
         AddExtension[] extensions = getAnnotations(testClass, AddExtension.class);
@@ -195,7 +195,7 @@ class HelidonJunitExtension implements BeforeAllCallback,
             ConfigMeta methodLevelConfigMeta = classLevelConfigMeta.nextMethod();
             methodLevelConfigMeta.addConfig(configs);
             methodLevelConfigMeta.configuration(method.getAnnotation(Configuration.class));
-            methodLevelConfigMeta.configBlob(method.getAnnotation(ConfigBlob.class));
+            methodLevelConfigMeta.addConfigBlock(method.getAnnotation(AddConfigBlock.class));
 
             configure(methodLevelConfigMeta);
 
@@ -324,7 +324,7 @@ class HelidonJunitExtension implements BeforeAllCallback,
                 }
             });
             if (configMeta.type != null) {
-                builder.withSources(configMeta.type.configSource(configMeta.content));
+                builder.withSources(configMeta.type.configSource(configMeta.block));
             }
             config = builder
                     .withSources(MpConfigSources.create(configMeta.additionalKeys))
@@ -607,7 +607,7 @@ class HelidonJunitExtension implements BeforeAllCallback,
         private final Map<String, String> additionalKeys = new HashMap<>();
         private final List<String> additionalSources = new ArrayList<>();
         private Type type;
-        private String content;
+        private String block;
         private boolean useExisting;
         private String profile;
 
@@ -640,12 +640,12 @@ class HelidonJunitExtension implements BeforeAllCallback,
             additionalKeys.put("mp.config.profile", profile);
         }
 
-        private void configBlob(ConfigBlob config) {
+        private void addConfigBlock(AddConfigBlock config) {
             if (config == null) {
                 return;
             }
             this.type = config.type();
-            this.content = config.content();
+            this.block = config.value();
         }
 
         ConfigMeta nextMethod() {
