+++ microprofile/tests/testing/testng/src/test/java/io/helidon/microprofile/tests/testing/testng/TestConfigBlobYaml.java
@@ -21,15 +21,15 @@ import static org.hamcrest.MatcherAssert.assertThat;
 
 import jakarta.inject.Inject;
 
-import io.helidon.microprofile.testing.testng.ConfigBlob;
-import io.helidon.microprofile.testing.testng.ConfigBlob.Type;
+import io.helidon.microprofile.testing.testng.AddConfigBlock;
+import io.helidon.microprofile.testing.testng.AddConfigBlock.Type;
 import io.helidon.microprofile.testing.testng.HelidonTest;
 
 import org.eclipse.microprofile.config.inject.ConfigProperty;
 import org.testng.annotations.Test;
 
 @HelidonTest
-@ConfigBlob(type = Type.YAML, content = """
+@AddConfigBlock(type = Type.YAML, value = """
     another1:
       key: "another1.value"
     another2:
